requirements.txt
# Core LLM and AI Libraries
langchain>=0.1.0
langgraph>=0.1.0
langfuse>=2.0.0
openai>=1.0.0
anthropic>=0.15.0
langchain-google-genai>=1.0.0

# Vector Store and Embeddings
chromadb>=0.4.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing and Evaluation
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Configuration and Logging
pyyaml>=6.0
loguru>=0.7.0

# Web Framework (if needed for API)
fastapi>=0.100.0
uvicorn>=0.22.0

# Jupyter for experimentation
jupyter>=1.0.0
ipykernel>=6.0.0

# Development tools
black>=23.0.0
flake8>=6.0.0
pre-commit>=3.0.0
